requirements.txt
fastapi
uvicorn
pdfminer.six
openai
python-multipart
